[irep2] Codex audit pass: delete dead helpers and stale transition comments
Apply the high-confidence cleanups from the post-flatten Codex review:

* Delete unused scaffold:
  - expr2t::expr_operands / Expr_operands typedefs (irep2.h);
  - esbmct::field_traits template (irep2.h) — K::fields supplanted it;
  - struct_union_member_pretty_names (decl + def) — zero callers;
  - #if 0 assertion block in index2t (irep2_expr.h).

* Inline single-call wrappers:
  - make_comp_expr — only caller in smt_memspace.cpp now switches inline;
  - get_vector_subtype — replaced with to_vector_type(t).subtype directly;
  - is_ieee_arith2_expr / is_bit2_expr — only caller in smt_conv.cpp's
    distribute_vector_operation now uses one switch covering ieee/arith/bit
    2-ops.

* Merge irep2_templates.h and irep2_template_utils.h into irep2_dispatch.h;
  both header files deleted. The unit test no longer needs to include them
  directly. Update README's file inventory.

* Strip transition-era / "@extends X_data" / "preserved verbatim" / "Flat-
  layout migration marker" / "old stringy irep" doc comments throughout
  irep2.h, irep2_expr.h, irep2_type.h. These describe history, not the
  current behaviour, and belong in the commit log rather than the headers.
